Software/requirements.txt
allensdk==2.13.6
matplotlib==3.4.2
numpy==1.21.5
opencv_python==4.6.0.66
pandas==1.4.2
pandas_profiling==3.5.0
paramiko==2.11.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2_binary==2.9.3
PyQt5==5.15.7
pyqtgraph==0.12.4
QtRangeSlider==0.1.5
scikit_learn==1.2.0
scipy==1.7.3
SimpleITK==2.2.1
visvis==1.13.0
xmltodict==0.13.0
